+++ tests/unit/client_tests.py
@@ -27,17 +27,19 @@ import base64
 import sys
 import os
 import datetime
+import mock
 
 from requests import ConnectTimeout, HTTPError
 from time import sleep
 
 from cloudant import cloudant, cloudant_bluemix, couchdb, couchdb_admin_party
 from cloudant.client import Cloudant, CouchDB
+from cloudant._client_session import BasicSession, CookieSession
+from cloudant.database import CloudantDatabase
 from cloudant.error import CloudantArgumentError, CloudantClientException
 from cloudant.feed import Feed, InfiniteFeed
-from cloudant._common_util import CookieSession
 
-from .unit_t_db_base import UnitTestDbBase
+from .unit_t_db_base import skip_if_not_cookie_auth, UnitTestDbBase
 from .. import bytes_, str_
 
 class CloudantClientExceptionTests(unittest.TestCase):
@@ -162,6 +164,7 @@ class ClientTests(UnitTestDbBase):
             self.client.disconnect()
             self.assertIsNone(self.client.r_session)
 
+    @skip_if_not_cookie_auth
     def test_auto_renew_enabled(self):
         """
         Test that CookieSession is used when auto_renew is enabled.
@@ -176,6 +179,7 @@ class ClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_auto_renew_enabled_with_auto_connect(self):
         """
         Test that CookieSession is used when auto_renew is enabled along with
@@ -190,6 +194,7 @@ class ClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_session(self):
         """
         Test getting session information.  
@@ -205,6 +210,7 @@ class ClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_session_cookie(self):
         """
         Test getting the session cookie.
@@ -219,6 +225,101 @@ class ClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @mock.patch('cloudant._client_session.Session.request')
+    def test_session_basic(self, m_req):
+        """
+        Test using basic access authentication.
+        """
+        m_response_ok = mock.MagicMock()
+        type(m_response_ok).status_code = mock.PropertyMock(return_value=200)
+        m_response_ok.json.return_value = ['animaldb']
+        m_req.return_value = m_response_ok
+
+        client = Cloudant('foo', 'bar', url=self.url, use_basic_auth=True)
+        client.connect()
+        self.assertIsInstance(client.r_session, BasicSession)
+
+        all_dbs = client.all_dbs()
+
+        m_req.assert_called_once_with(
+            'GET',
+            self.url + '/_all_dbs',
+            allow_redirects=True,
+            auth=('foo', 'bar'),  # uses HTTP Basic Auth
+            timeout=None
+        )
+
+        self.assertEquals(all_dbs, ['animaldb'])
+
+    @mock.patch('cloudant._client_session.Session.request')
+    def test_session_basic_with_no_credentials(self, m_req):
+        """
+        Test using basic access authentication with no credentials.
+        """
+        m_response_ok = mock.MagicMock()
+        type(m_response_ok).status_code = mock.PropertyMock(return_value=200)
+        m_req.return_value = m_response_ok
+
+        client = Cloudant(None, None, url=self.url, use_basic_auth=True)
+        client.connect()
+        self.assertIsInstance(client.r_session, BasicSession)
+
+        db = client['animaldb']
+
+        m_req.assert_called_once_with(
+            'HEAD',
+            self.url + '/animaldb',
+            allow_redirects=False,
+            auth=None,  # ensure no authentication specified
+            timeout=None
+        )
+
+        self.assertIsInstance(db, CloudantDatabase)
+
+    @mock.patch('cloudant._client_session.Session.request')
+    def test_change_credentials_basic(self, m_req):
+        """
+        Test changing credentials when using basic access authentication.
+        """
+        # mock 200
+        m_response_ok = mock.MagicMock()
+        m_response_ok.json.return_value = ['animaldb']
+
+        # mock 401
+        m_response_bad = mock.MagicMock()
+        m_response_bad.raise_for_status.side_effect = HTTPError('401 Unauthorized')
+
+        m_req.side_effect = [m_response_bad, m_response_ok]
+
+        client = Cloudant('foo', 'bar', url=self.url, use_basic_auth=True)
+        client.connect()
+        self.assertIsInstance(client.r_session, BasicSession)
+
+        with self.assertRaises(HTTPError):
+            client.all_dbs()  # expected 401
+
+        m_req.assert_called_with(
+            'GET',
+            self.url + '/_all_dbs',
+            allow_redirects=True,
+            auth=('foo', 'bar'),  # uses HTTP Basic Auth
+            timeout=None
+        )
+
+        # use valid credentials
+        client.change_credentials('baz', 'qux')
+        all_dbs = client.all_dbs()
+
+        m_req.assert_called_with(
+            'GET',
+            self.url + '/_all_dbs',
+            allow_redirects=True,
+            auth=('baz', 'qux'),  # uses HTTP Basic Auth
+            timeout=None
+        )
+        self.assertEquals(all_dbs, ['animaldb'])
+
+    @skip_if_not_cookie_auth
     def test_basic_auth_str(self):
         """
         Test getting the basic authentication string.
@@ -493,6 +594,7 @@ class CloudantClientTests(UnitTestDbBase):
     Cloudant specific client unit tests
     """
 
+    @skip_if_not_cookie_auth
     def test_cloudant_session_login(self):
         """
         Test that the Cloudant client session successfully authenticates.
@@ -505,6 +607,7 @@ class CloudantClientTests(UnitTestDbBase):
         self.client.session_login()
         self.assertNotEqual(self.client.session_cookie(), old_cookie)
 
+    @skip_if_not_cookie_auth
     def test_cloudant_session_login_with_new_credentials(self):
         """
         Test that the Cloudant client session fails to authenticate when
@@ -517,6 +620,7 @@ class CloudantClientTests(UnitTestDbBase):
 
         self.assertTrue(str(cm.exception).find('Name or password is incorrect'))
 
+    @skip_if_not_cookie_auth
     def test_cloudant_context_helper(self):
         """
         Test that the cloudant context helper works as expected.
@@ -528,6 +632,7 @@ class CloudantClientTests(UnitTestDbBase):
         except Exception as err:
             self.fail('Exception {0} was raised.'.format(str(err)))
 
+    @skip_if_not_cookie_auth
     def test_cloudant_bluemix_context_helper(self):
         """
         Test that the cloudant_bluemix context helper works as expected.
@@ -592,6 +697,7 @@ class CloudantClientTests(UnitTestDbBase):
             'https://{0}.cloudant.com'.format(self.account)
             )
 
+    @skip_if_not_cookie_auth
     def test_bluemix_constructor(self):
         """
         Test instantiating a client object using a VCAP_SERVICES environment
@@ -624,6 +730,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             c.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_bluemix_constructor_specify_instance_name(self):
         """
         Test instantiating a client object using a VCAP_SERVICES environment
@@ -656,6 +763,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             c.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_bluemix_constructor_with_multiple_services(self):
         """
         Test instantiating a client object using a VCAP_SERVICES environment
@@ -723,6 +831,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_connect_timeout(self):
         """
         Test that a connect timeout occurs when instantiating
@@ -749,6 +858,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_billing_data(self):
         """
         Test the retrieval of billing data
@@ -843,6 +953,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_volume_usage_data(self):
         """
         Test the retrieval of volume usage data
@@ -934,6 +1045,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_requests_usage_data(self):
         """
         Test the retrieval of requests usage data
@@ -1025,6 +1137,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_shared_databases(self):
         """
         Test the retrieval of shared database list
@@ -1035,6 +1148,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_generate_api_key(self):
         """
         Test the generation of an API key for this client account
@@ -1048,6 +1162,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_cors_configuration(self):
         """
         Test the retrieval of the current CORS configuration for this client
@@ -1061,6 +1176,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_cors_origins(self):
         """
         Test the retrieval of the CORS origins list
@@ -1072,6 +1188,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_disable_cors(self):
         """
         Test disabling CORS (assuming CORS is enabled)
@@ -1092,6 +1209,7 @@ class CloudantClientTests(UnitTestDbBase):
         finally:
             self.client.disconnect()
 
+    @skip_if_not_cookie_auth
     def test_update_cors_configuration(self):
         """
         Test updating CORS configuration